backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
pydantic==2.10.3
uvicorn==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.0.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
mangum==0.17.0